Added 2D probing detection method
It generates rotations to check with the expensive 2D matching stage
this can account for differences in translation just fine

By default, it is much slower than existing 2D search
However, it has more potential and can more reliably detect targets with few points

By redistributing the rotations it is also possible to use e.g. the IMU to speed it up
This has been implemented partially as a proof of concept, but not used or tested

This method is required for some targets, but is a step down for others.